About and Contact Us Page
Completed the layout of the about page with relevant information.
Completed the layout of the contact us page with relevant information.
Added a form on the contact us page for clients to submit queries, that is then saved to the database.
Added functionality for a sticky header (this functionality is currently not being used).
Added a back to top function to the footer.
*Fixed the navigation dropdown issue where it was not displaying the correct brands.
Added the navigation to all pages.