Cart Functionality
Added functionality that allows an user to add items to their cart.
Added functionality that displays the items in the users cart when navigating to the cart page.
Added functionality to display a message when cart is empty.\
Added functionality that allows the user to remove items from their cart.
Added functionality that totals the different items prices and displays a subtotal.
Added functionality to display the amount of items in an user's cart on all pages.