Images
Added product images to the database.
Added functionality that displays product images.
Added functionality to display an image if it is hovered over on the product information page.
Added functionality that updates the cart subtotal when a products quantity changes.
Added functionality open a new tab when a user clicks on socials.