Image Subtle Zoom
Added functionality to the api to give a response if an entered email already exists on the database (used then users attempt to create or update their account).
Added functionality to display an error when no products are found to display on the product page.
Removed a meta tag that caused the mobile view to respond incorrectly.
Added a subtle zoom-in effect on the main image of the product information screen.