Cart Page Updating
Added functionality to update the users cart when adding and removing items from their cart. This data will stay persistent even after the user leaves the page.
Changed a number input to tel.
Added fail safe functionality on the product page for when the navigation dropdown does not load correctly.  This behavior has on been experienced on the product page.